Removing delete.go and moving its functions to storage.go
